feat: make all 6 SDKs fully standalone with independent crypto
- Python: Add standalone crypto, ratchet, X3DH, session, safety_number, group
  Pure Python with cryptography lib. No Rust FFI needed.
- JavaScript: Add standalone crypto (noble libs), ratchet, X3DH, session,
  safety_number, group. Remove @noble/ed25519 dependency.
- Go: Add standalone crypto, ratchet, X3DH, session, safety_number, group
  with golang.org/x/crypto.
- Dart: Replace FFI wrapper with pure Dart crypto (cryptography package).
  Add ratchet, X3DH, session modules.
- C++: Add standalone X3DH and DoubleRatchet modules with OpenSSL.
  Update session, group, safety_number.
- Java: Add standalone GroupSession, SafetyNumber, VerificationQrCode.
  Fix DoubleRatchet KDF and X3DH transcript binding.

Cross-SDK verification:
- JS safety number matches Python exactly (26/26 tests pass)
- JS X3DH + Double Ratchet roundtrip verified (A→B, B→A, interleaved)
- Wire format verified: dh_pub(32) + msg_num(4) + nonce(12) + ct + tag(16)

Package updates:
- Python: setup.py updated to pure Python, version 3.0.1
- CI: Remove build-native dependency from Dart and C++ SDK jobs
- All protocol constants match: MIN_COMPATIBLE_VERSION=9, MAX_CHAIN_MESSAGES=4000

+++ sdks/cpp/include/sibna/types.hpp
@@ -173,7 +173,7 @@ struct SessionInfo {
 struct GroupInfo {
     group_id id;
     size_t member_count = 0;
-    uint64_t epoch = 0;
+    uint32_t epoch = 0;  // uint32_t per CRITICAL requirement #4
     std::optional<std::chrono::system_clock::time_point> created_at;
     std::optional<std::chrono::system_clock::time_point> last_activity;
 };
